+++ Package.swift
@@ -8,7 +8,7 @@ let package = Package(
   defaultLocalization: "en",
   platforms: [
     .iOS(.v13),
-    .macOS(.v10_12),
+    .macOS(.v10_15),
     .watchOS("6.2")
   ],
   products: [